Fonction bissextille

+++ bissextile.cpp
@@ -1,6 +1,6 @@
-#include "bissextiles.h"
+#include "bissextile.h"
 
-bool testBissextile(unsigned int annee){
+bool testBissextile(unsigned short annee){
    if(annee%400==0 || (annee%4==0 && annee%100!=0)){
       return true;
    }
